backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
sqlalchemy==2.0.35
alembic==1.13.2
asyncpg==0.29.0
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
argon2-cffi==23.1.0
email-validator==2.2.0
